src-py/pyproject.toml
[project]
name = "dcp-cu"
version = "0.1.0"
description = "Data capitalism project credit union"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pywebview==6.1",
    "pyinstaller",
    "setuptools",
    "pydantic",
    "pydantic-to-typescript",
    "pysubs2",
    "charset_normalizer",
    "pandas",
    "openpyxl",
    "watchdog",
]
